create GObject trait and move connect method to there
