+++ src/glib/ffi.rs
@@ -154,6 +154,11 @@ extern "C" {
     pub fn g_object_ref(object: *mut C_GObject) -> *mut C_GObject;
     pub fn g_object_unref(object: *mut C_GObject);
 
+    pub fn glue_signal_connect(g_object: *mut C_GObject,
+                               signal: *const c_char,
+                               func: Option<extern "C" fn()>,
+                               user_data: *const c_void);
+
     //=========================================================================
     // GType constants
     //=========================================================================
