+++ src/gtk/traits/widget.rs
@@ -15,7 +15,6 @@
 
 use libc::{c_int, c_char};
 use gtk::ffi;
-use gtk::signals::Signal;
 use gtk::enums;
 use gdk;
 use gtk;
@@ -664,26 +663,6 @@ pub trait Widget: ffi::FFIWidget {
     fn get_margin_bottom(&mut self) -> i32 {
         unsafe {
             ffi::gtk_widget_get_margin_bottom(self.get_widget()) as i32
-        }
-    }
-
-    fn connect<'a>(&self, signal: Box<Signal<'a>>) -> () {
-        use std::mem::transmute;
-
-        unsafe {
-            let signal_name     = signal.get_signal_name().to_string();
-            let trampoline      = signal.get_trampoline();
-
-            let user_data_ptr   = transmute(box signal);
-
-            signal_name.replace("_", "-").with_c_str(|signal_name| {
-                ffi::glue_signal_connect(
-                    self.get_widget(),
-                    signal_name,
-                    Some(trampoline),
-                    user_data_ptr
-                )
-            });
         }
     }
